# Define most of Pathname in Ruby (redo)