--- a/lib/pathname.rb
+++ b/lib/pathname.rb
@@ -252,10 +252,12 @@ class Pathname
   alias === ==
   alias eql? ==
 
-  # Provides for comparing pathnames, case-sensitively.
-  def <=>(other)
-    return nil unless Pathname === other
-    @path.tr('/', "\0") <=> other.to_s.tr('/', "\0")
+  unless method_defined?(:<=>, false)
+    # Provides for comparing pathnames, case-sensitively.
+    def <=>(other)
+      return nil unless Pathname === other
+      @path.tr('/', "\0") <=> other.to_s.tr('/', "\0")
+    end
   end
 
   def hash # :nodoc: