… .github/workflows/test.yml
… test/pathname/test_pathname.rb