@eregon 2025-07-18
@hsbt Let's discuss your concerns and suggestions here.
You said in https://bugs.ruby-lang.org/issues/17473#note-27
> Please separate the small PRs. I want to reduce the side effect like https://github.com/ruby/ruby/pull/13906.

Can you make a concrete suggestion by what you mean by small PRs for this change?

I could make a PR with fewer commits, but every commit until [Handle Windows NTFS edge case in Pathname#sub_ext](https://github.com/ruby/pathname/pull/57/commits/99eaa8c339b8bdb266bc8662d2f9e5fbf827148b) is strictly necessary, otherwise the CI doesn't pass.
That leaves only [Optimize Pathname#initialize to avoid extra __send__](https://github.com/ruby/pathname/pull/57/commits/635d8a76d6069cf45621edcf5d052431e28a5a2e) and [Optimize Pathname#initialize to avoid extra ivar accesses](https://github.com/ruby/pathname/pull/57/commits/91d6fa6dfa300b52466fafe920ba9d04a8da4f4a) which are trivial, and then commits to address @byroot's review.

If you are asking a smaller diff in general I think that is not feasible, e.g. making a PR per method would take months of work and still be the exact same end result. The approach here as detailed in the first commit message, [Restore lib/pathname.rb from ext/pathname/lib/pathname.rb at ed9270a](https://github.com/ruby/pathname/pull/57/commits/b26ed2822f86c925e54e87bc5b66cebd787684e8) is to use the Ruby code of pathname.rb from before the translation to C. There is no meaningful way to break that in smaller changes. And that code has already been reviewed, it was exactly the code in Pathname before the translation to C started.

Please take the time to read the commit messages, they should make it very clear what I did and what needs deeper review (e.g. imported code from the gem as-is doesn't).
Just browsing through the commit messages should also make it clear I took great care to have a very clear git history of the changes here with not a single extra line of diff.