@eregon 2025-08-05
I also updated the description to add a table summarizing the performance gains, and filed https://bugs.ruby-lang.org/issues/21532 to make a ticket specifically about this, since IIUC that's what @hsbt wants.