@headius 2025-08-21
Huge +1 from the JRuby side. Pathname moving to C was unnecessary then and a hassle for all of us now (including CRuby due to JIT+C interaction. Ship it!

I'll direct any additional comments to the ruby-lang issue.