@eregon 2025-08-21
Approved by @akr in https://bugs.ruby-lang.org/issues/21532#note-5, merging :tada: