@eregon 2025-08-21
@headius Thanks for the support, should have pinged you earlier :)

> Ok, I went to the ruby-lang issue but the last couple of comments directed me back here.

I'm not sure from which issue your points come from, is it https://bugs.ruby-lang.org/issues/17473 maybe?

> Pathname becoming core, what happens to the gem?

I believe we need to keep it, we need to keep the gem for older Ruby versions anyway.
Also this gem [replaces the ::Pathname constant if any when loaded](https://github.com/ruby/pathname/blob/658648cc955a3e5284561a28d745b119d4d3ea62/lib/pathname.rb#L13-L29), which seems good if e.g. there would be a bug fix in the gem without a CRuby release.

> Pathname becoming core, how can it be pure-Ruby?

Some parts of Pathname were already Ruby, see e.g. https://github.com/ruby/ruby/blob/master/pathname_builtin.rb and `lib/pathname.rb` before this PR. So not a problem in any case.